Display correct but non-functional components on new desktop page
